Move swap-domain hooks to appropriate folder
