+++ src/cow-react/pages/NewSwap/index.tsx
@@ -9,12 +9,12 @@ import {
   useUnknownImpactWarning,
 } from 'state/swap/hooks'
 import { Field } from 'state/swap/actions'
-import { useSetupSwapState } from 'cow-react/pages/NewSwap/hooks/useSetupSwapState'
+import { useSetupSwapState } from 'cow-react/swap/hooks/useSetupSwapState'
 import { useCurrencyBalance } from 'state/connection/hooks'
 import { useWeb3React } from '@web3-react/core'
 import { CurrencyInfo, SwapFormProps } from 'cow-react/pages/NewSwap/typings'
 import { useHigherUSDValue } from 'hooks/useStablecoinPrice'
-import { useSwapCurrenciesAmounts } from 'cow-react/pages/NewSwap/hooks/useSwapCurrenciesAmounts'
+import { useSwapCurrenciesAmounts } from 'cow-react/swap/hooks/useSwapCurrenciesAmounts'
 import usePriceImpact from 'hooks/usePriceImpact'
 import { useWrapType, WrapType } from 'hooks/useWrapCallback'
 import {
@@ -22,7 +22,7 @@ import {
   getOutputReceiveAmountInfo,
 } from 'cow-react/pages/NewSwap/helpers/tradeReceiveAmount'
 import { SwapButton } from 'pages/Swap/components/SwapButton/SwapButton'
-import { useSwapButtonContext } from 'pages/Swap/hooks/useSwapButtonContext'
+import { useSwapButtonContext } from '@src/cow-react/swap/hooks/useSwapButtonContext'
 import { useModalIsOpen } from 'state/application/hooks'
 import { ApplicationModal } from 'state/application/reducer'
 import { NewSwapModals, NewSwapModalsProps } from 'cow-react/pages/NewSwap/modals'
@@ -40,9 +40,9 @@ import { useIsSwapUnsupported } from 'hooks/useIsSwapUnsupported'
 import { useExpertModeManager, useUserSlippageTolerance } from 'state/user/hooks'
 import useCowBalanceAndSubsidy from 'hooks/useCowBalanceAndSubsidy'
 import { SwapForm } from 'cow-react/pages/NewSwap/pureComponents/SwapForm'
-import { useShowRecipientControls } from 'cow-react/pages/NewSwap/hooks/useShowRecipientControls'
+import { useShowRecipientControls } from 'cow-react/swap/hooks/useShowRecipientControls'
 import { TradeRates, TradeRatesProps } from 'cow-react/pages/NewSwap/pureComponents/TradeRates'
-import { useTradePricesUpdate } from 'cow-react/pages/NewSwap/hooks/useTradePricesUpdate'
+import { useTradePricesUpdate } from 'cow-react/swap/hooks/useTradePricesUpdate'
 import { tokenViewAmount } from 'cow-react/pages/NewSwap/helpers/tokenViewAmount'
 
 export function NewSwapPage() {
