+++ src/cow-react/swap/hooks/useHandleSwap.ts
@@ -1,12 +1,12 @@
 import { useCallback } from 'react'
-import confirmPriceImpactWithoutFee from '@src/components/swap/confirmPriceImpactWithoutFee'
+import confirmPriceImpactWithoutFee from 'components/swap/confirmPriceImpactWithoutFee'
 import { signSwapAnalytics, swapAnalytics } from 'utils/analytics'
 import { getProviderErrorMessage, isRejectRequestProviderError } from 'utils/misc'
 import { SwapCallbackParams, useSwapCallback } from 'hooks/useSwapCallback'
 import { Percent } from '@uniswap/sdk-core'
 import TradeGp from 'state/swap/TradeGp'
 import { useCloseModals } from 'state/application/hooks'
-import { useSwapConfirmManager } from 'pages/Swap/hooks/useSwapConfirmManager'
+import { useSwapConfirmManager } from 'cow-react/swap/hooks/useSwapConfirmManager'
 
 export interface HandleSwapInput {
   trade: TradeGp | undefined
