+++ src/cow-react/swap/hooks/useSwapButtonContext.ts
@@ -4,7 +4,7 @@ import { useDerivedSwapInfo, useDetectNativeToken, useSwapState } from 'state/sw
 import useTransactionDeadline from 'hooks/useTransactionDeadline'
 import { useExpertModeManager } from 'state/user/hooks'
 import { useCloseModals, useToggleWalletModal } from 'state/application/hooks'
-import { useSwapConfirmManager } from 'pages/Swap/hooks/useSwapConfirmManager'
+import { useSwapConfirmManager } from 'cow-react/swap/hooks/useSwapConfirmManager'
 import { Field } from 'state/swap/actions'
 import { TradeType } from '@uniswap/sdk-core'
 import { computeSlippageAdjustedAmounts } from 'utils/prices'
@@ -20,7 +20,7 @@ import { getSwapButtonState } from 'pages/Swap/helpers/getSwapButtonState'
 import { SwapButtonContext } from 'pages/Swap/components/SwapButton/SwapButton'
 import { useGetQuoteAndStatus } from 'state/price/hooks'
 import { OperationType } from 'components/TransactionConfirmationModal'
-import { useTransactionConfirmModal } from 'pages/Swap/hooks/useTransactionConfirmModal'
+import { useTransactionConfirmModal } from 'cow-react/swap/hooks/useTransactionConfirmModal'
 import { useSwapFlowContext } from 'pages/Swap/swapFlow/useSwapFlowContext'
 import { PriceImpact } from 'hooks/usePriceImpact'
