+++ src/custom/components/swap/EthFlow/index.tsx
@@ -31,10 +31,10 @@ import { GpModal } from 'components/Modal'
 import { ApprovalState } from 'hooks/useApproveCallback'
 import { useWrapCallback } from 'hooks/useWrapCallback'
 import { useDerivedSwapInfo, useDetectNativeToken, useSwapActionHandlers } from 'state/swap/hooks'
-import { useSwapConfirmManager } from 'pages/Swap/hooks/useSwapConfirmManager'
+import { useSwapConfirmManager } from '@src/cow-react/swap/hooks/useSwapConfirmManager'
 import { Field } from '@src/state/swap/actions'
 import { WRAPPED_NATIVE_CURRENCY } from 'constants/tokens'
-import { HandleSwapCallback } from 'pages/Swap/hooks/useHandleSwap'
+import { HandleSwapCallback } from '@src/cow-react/swap/hooks/useHandleSwap'
 
 const EthFlowModalContent = styled(ConfirmationModalContent)`
   padding: 22px;
