+++ src/custom/pages/Swap/swapFlow/useSwapFlowContext.ts
@@ -11,7 +11,7 @@ import { useDispatch } from 'react-redux'
 import { AppDispatch } from 'state'
 import { SwapFlowAnalyticsContext } from 'pages/Swap/swapFlow/steps/analytics'
 import useENSAddress from 'hooks/useENSAddress'
-import { useSwapConfirmManager } from 'pages/Swap/hooks/useSwapConfirmManager'
+import { useSwapConfirmManager } from '@src/cow-react/swap/hooks/useSwapConfirmManager'
 import { useWETHContract } from 'hooks/useContract'
 import { computeSlippageAdjustedAmounts } from 'utils/prices'
 import { PostOrderParams } from 'utils/trade'
